[MISC] Apply type erasure on the alphabet types to reduce the  compatibility overhead.
By applying this type erasure, we are actually generic to allow any user input as long as it fulfils the respective concepts
that are modeled explicitly for alphabets and scoring schemes.
It also is an initial step to reduce the boiler plate overloads in
seqan2 when we start adapting the algorithms step by step.